forward simulation of a diploid population
Forward simulation of a diploid population allowing for random sweepstakes and recurrent bottlenecks and  selection but without recombination
or population structure.   The population model is  a model of random sweepstakes where  in any given generation diploid individuals are randomly paired
and each pair independently  produces a random  number of juveniles, or potential offspring, according to a given distribution.  The  law on the number of juveniles 
can be heavy-tailed modeling random sweepstakes, or  light-tailed mimicking  Wright-Fisher dynamics.   Bottlenecks  occur with a given  probability in any given generation,  and  when a bottleneck occurs the population is reduced to a given number of individuals.   The remaining individuals  then  produce juveniles. If the total number of juveniles
is smaller than some given number, the  carrying capacity of the environment, all  juveniles survive.  If the  total number of juveniles exceeds the carrying capacity selection kicks in and we sort the juveniles based on the  average  weight of the mutations  carried by the juveniles.    The weight of the mutations are  sampled  from a shifted gamma distribution.